refactor: archive 페이지 api 제거 및 json 파일로 교체
