+++ src/apis/getProducts.ts
@@ -1,25 +1,24 @@
 export interface ProductItemResponse {
   productId: number
+  type: string[]
   name: string
   introduction: string
   thumbnailUrl: string
   siteUrl: string | null
 }
 
-export const getProducts = async (type: string): Promise<ProductItemResponse[]> => {
-  const convertType = String(type).toLowerCase()
-  const response = await (
-    await fetch(`${process.env.NEXT_PUBLIC_API_URL}/products/${convertType}`, {
-      method: 'GET',
-      headers: {
-        'Content-Type': 'application/json',
-      },
-    })
-  ).json()
+export const getProducts = async (): Promise<ProductItemResponse[]> => {
+  try {
+    const response = await fetch(`${process.env.NEXT_PUBLIC_FETCH_URL}/data/products.json`)
+    const data = await response.json()
 
-  if (!response.isSuccess) {
-    throw new Error(response.message)
+    if (!Array.isArray(data)) {
+      throw new Error('Invalid data format: Expected an array')
+    }
+
+    return data
+  } catch (error) {
+    console.error('Failed to fetch products:', error)
+    throw error
   }
-
-  return response.payload
 }
